+++ pkg/handlers/backup/state_uploaderror.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -30,7 +30,8 @@ func stateUploadErrorHandler(h *handler, backup *backupApi.ArangoBackup) (*backu
 	// no more retries - move to failed state
 	if !backup.Status.Backoff.ShouldBackoff(backup.Spec.Backoff) {
 		return wrapUpdateStatus(backup,
-			updateStatusState(backupApi.ArangoBackupStateFailed, "out of Upload retries"),
+			updateStatusStateOnly(backupApi.ArangoBackupStateFailed),
+			cleanBackOff(),
 			cleanStatusJob())
 	}