+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Bugfix) Check Connection to the ArangoDB before creating Backup
 - (Feature) Deployment & Members Condition metrics
 - (Maintenance) Update Go to 1.21.8 & Dependencies
+- (Feature) (ArangoBackup) Propagate message during retries
 
 ## [1.2.38](https://github.com/arangodb/kube-arangodb/tree/1.2.38) (2024-02-22)
 - (Feature) Extract GRPC Server